+++ noodles-bam/src/reader.rs
@@ -10,6 +10,7 @@ pub use self::{query::Query, records::Records, unmapped_records::UnmappedRecords
 use std::{
     ffi::CStr,
     io::{self, Read, Seek},
+    mem,
 };
 
 use byteorder::{LittleEndian, ReadBytesExt};
@@ -377,7 +378,7 @@ where
     ReferenceSequence::new(name, l_ref).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
 }
 
-fn read_record<R>(reader: &mut R, record: &mut Record) -> io::Result<usize>
+pub(crate) fn read_record<R>(reader: &mut R, record: &mut Record) -> io::Result<usize>
 where
     R: Read,
 {
@@ -387,8 +388,49 @@ where
         Err(e) => return Err(e),
     };
 
-    record.resize(block_size);
-    reader.read_exact(record)?;
+    record.ref_id = reader.read_i32::<LittleEndian>()?;
+    record.pos = reader.read_i32::<LittleEndian>()?;
+
+    let l_read_name = reader.read_u8().and_then(|n| {
+        usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
+    })?;
+
+    record.mapq = reader.read_u8()?;
+    record.bin = reader.read_u16::<LittleEndian>()?;
+
+    let n_cigar_op = reader.read_u16::<LittleEndian>().and_then(|n| {
+        usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
+    })?;
+
+    record.flag = reader.read_u16::<LittleEndian>()?;
+
+    let l_seq = reader.read_u32::<LittleEndian>().and_then(|n| {
+        usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
+    })?;
+    record.l_seq = l_seq;
+
+    record.next_ref_id = reader.read_i32::<LittleEndian>()?;
+    record.next_pos = reader.read_i32::<LittleEndian>()?;
+    record.tlen = reader.read_i32::<LittleEndian>()?;
+
+    record.read_name.resize(l_read_name, Default::default());
+    reader.read_exact(&mut record.read_name)?;
+
+    let cigar_len = n_cigar_op * mem::size_of::<u32>();
+    record.cigar.resize(cigar_len, Default::default());
+    reader.read_exact(&mut record.cigar)?;
+
+    let seq_len = (l_seq + 1) / 2;
+    record.seq.resize(seq_len, Default::default());
+    reader.read_exact(&mut record.seq)?;
+
+    record.qual.resize(l_seq, Default::default());
+    reader.read_exact(&mut record.qual)?;
+
+    let data_offset = 32 + l_read_name + cigar_len + seq_len + l_seq;
+    let data_len = block_size - data_offset;
+    record.data.resize(data_len, Default::default());
+    reader.read_exact(&mut record.data)?;
 
     Ok(block_size)
 }
@@ -496,16 +538,27 @@ mod tests {
     #[test]
     fn test_read_record() -> io::Result<()> {
         let data = [
-            0x04, 0x00, 0x00, 0x00, // block_size = 4
-            0x6e, 0x64, 0x6c, 0x73, // ...
+            0x22, 0x00, 0x00, 0x00, // block_size = 34
+            0xff, 0xff, 0xff, 0xff, // ref_id = -1
+            0xff, 0xff, 0xff, 0xff, // pos = -1
+            0x02, // l_read_name = 2
+            0xff, // mapq = 255
+            0x48, 0x12, // bin = 4680
+            0x00, 0x00, // n_cigar_op = 0
+            0x04, 0x00, // flag = 4
+            0x00, 0x00, 0x00, 0x00, // l_seq = 0
+            0xff, 0xff, 0xff, 0xff, // next_ref_id = -1
+            0xff, 0xff, 0xff, 0xff, // next_pos = -1
+            0x00, 0x00, 0x00, 0x00, // tlen = 0
+            0x2a, 0x00, // read_name = "*\x00"
         ];
 
         let mut reader = &data[..];
         let mut record = Record::default();
         let block_size = read_record(&mut reader, &mut record)?;
 
-        assert_eq!(block_size, 4);
-        assert_eq!(&record[..], &data[4..]);
+        assert_eq!(block_size, 34);
+        assert_eq!(record, Record::default());
 
         Ok(())
     }